+++ rules-tests/CodingStyle/Rector/FuncCall/CallUserFuncArrayToVariadicRector/config/configured_rule.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-use Rector\CodingStyle\Rector\FuncCall\CallUserFuncCallToVariadicRector;
+use Rector\CodingStyle\Rector\FuncCall\CallUserFuncArrayToVariadicRector;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
 
 return static function (ContainerConfigurator $containerConfigurator): void {
     $services = $containerConfigurator->services();
-    $services->set(CallUserFuncCallToVariadicRector::class);
+    $services->set(CallUserFuncArrayToVariadicRector::class);
 };
